Add files via upload
Statistics is a collection of mathematics to deal with data organization, analysis, and interpretation. Three main statistical methods are used in the data analysis: descriptive statistics, inferential statistics, and regressions analysis.

Descriptive statistics summarize the data and usually focuses on the distribution (normal distribution or binomial distribution), the central tendency (it describes the data with respect to the center of the data, and it can be the mean, median, and mode of the data), and the dispersion of data (it describes the spread of the data, and it can be the variance, standard deviation, and interquartile range).